Added json parser for twexted text Added preview of example on page load Index file is now HTML, php is not used or needed
The json parser is json2.js from http://json.org/js.html

Javascript cannot write to file unless its a cookie, so we need to take a "smart" different approach to this
to keep it 100% client side javascript only. Saving on server side can be possible by sending
the contents of twexted javascript variable to a server by ajax request.